refactor: ajuste de type e bugs

+++ source/openmp.c
@@ -5,9 +5,9 @@ void mandelbrot_openmp(unsigned char *buffer, const Config *cfg){
 
     #pragma omp parallel for num_threads(cfg->q_threads) schedule(static)
 
-    for (long lin = 0; lin < cfg->altura; lin++){
+    for (int lin = 0; lin < cfg->altura; lin++){
         double ci = IMAG_MIN + lin * (IMAG_MAX - (IMAG_MIN)) / cfg->altura;
-        for (long col = 0; col < cfg->largura; col++) {
+        for (int col = 0; col < cfg->largura; col++) {
             double cr = REAL_MIN + col * (REAL_MAX - (REAL_MIN)) / cfg->largura;
 
             int iter = mandelbrot(cr, ci, cfg->max_iter);
